+++ src/Http/Controllers/Elements/SaveElementController.php
@@ -10,12 +10,14 @@ use CraftCms\Cms\Element\Drafts;
 use CraftCms\Cms\Element\ElementActivity;
 use CraftCms\Cms\Element\Elements;
 use CraftCms\Cms\Element\Enums\ElementActivityType;
+use CraftCms\Cms\Element\Exceptions\InvalidElementException;
 use CraftCms\Cms\Element\Exceptions\UnsupportedSiteException;
 use CraftCms\Cms\Element\Validation\ElementRules;
 use CraftCms\Cms\Http\Controllers\Elements\Concerns\SavesElement;
 use CraftCms\Cms\Http\Requests\ElementRequest;
 use CraftCms\Cms\Http\Responses\ElementResponse;
 use CraftCms\Cms\Site\Sites;
+use CraftCms\Cms\Workflow\Workflows;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
@@ -35,6 +37,7 @@ readonly class SaveElementController
         private ElementActivity $elementActivity,
         private Elements $elements,
         private Sites $sites,
+        private Workflows $workflows,
     ) {}
 
     public function store(): Response
@@ -62,6 +65,7 @@ readonly class SaveElementController
         }
 
         $isNotNew = $element->id;
+        $saveAsDraft = $this->workflows->requiresApproval($element);
         if ($isNotNew) {
             $mutex = Cache::lock("element:$element->id", 15);
             if (! $mutex->get()) {
@@ -75,15 +79,25 @@ readonly class SaveElementController
 
         try {
             $namespace = $this->request->header('X-Craft-Namespace');
-            // crossSiteValidate only if it's multisite, element supports drafts and we're not in a slideout
-            $success = $this->elements->saveElement(
-                $element,
-                crossSiteValidate: (
-                    $namespace === null
-                    && $this->sites->isMultiSite()
-                    && Gate::check('createDrafts', $element)
-                ),
-            );
+            if ($saveAsDraft && $isNotNew) {
+                $element = $this->drafts->createDraft($element, $this->request->craftUser()?->getCraftUserId());
+                $success = true;
+            } elseif ($saveAsDraft) {
+                $success = $this->drafts->saveElementAsDraft($element, $this->request->craftUser()?->getCraftUserId());
+            } else {
+                // crossSiteValidate only if it's multisite, element supports drafts and we're not in a slideout
+                $success = $this->elements->saveElement(
+                    $element,
+                    crossSiteValidate: (
+                        $namespace === null
+                        && $this->sites->isMultiSite()
+                        && Gate::check('createDrafts', $element)
+                    ),
+                );
+            }
+        } catch (InvalidElementException $e) {
+            $element = $e->element;
+            $success = false;
         } catch (UnsupportedSiteException $e) {
             $element->errors()->add('siteId', $e->getMessage());
             $success = false;
@@ -102,7 +116,7 @@ readonly class SaveElementController
         $this->elementActivity->trackActivity($element, ElementActivityType::Save);
 
         // See if the user happens to have a provisional element. If so delete it.
-        $provisional = $element::find()
+        $provisional = $saveAsDraft ? null : $element::find()
             ->provisionalDrafts()
             ->draftOf($element->id)
             ->draftCreator($this->request->craftUser()?->asElement())