+++ src/Providers/CraftServiceProvider.php
@@ -29,6 +29,8 @@ use CraftCms\Cms\Twig\TwigServiceProvider;
 use CraftCms\Cms\Update\UpdatesServiceProvider;
 use CraftCms\Cms\User\UserServiceProvider;
 use CraftCms\Cms\View\ViewServiceProvider;
+use CraftCms\Cms\Workflow\UserReview\UserReviewServiceProvider;
+use CraftCms\Cms\Workflow\WorkflowServiceProvider;
 use Illuminate\Support\AggregateServiceProvider;
 use Override;
 
@@ -61,6 +63,8 @@ class CraftServiceProvider extends AggregateServiceProvider
         FieldLayoutServiceProvider::class,
         SectionServiceProvider::class,
         EntryServiceProvider::class,
+        WorkflowServiceProvider::class,
+        UserReviewServiceProvider::class,
         StructureServiceProvider::class,
         QueueServiceProvider::class,
     ];